api/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
alembic==1.18.4
python-dotenv==1.2.2
ryanair-py==2.3.0
apscheduler==3.11.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.5.0

email-validator==2.2.0
requests==2.32.3

# testing
pytest==9.0.3
httpx==0.28.1
pytest-cov==6.1.0
